refactor(tui): route render.rs through the ViewModel only (#289)
docs/agents/architecture.md says paint helpers "apply theme/layout only —
no business rules, FS, or network," reading only from the pure ViewModel.
render.rs defined 5 functions taking state: &AppState and calling real
AppState business methods (diff_title, diff_footer, confirm_prompt,
confirm_modal_style, gist_row_display) — a seam violation on paper. In
practice none of them are ever called from real paint code (functions
taking &mut Frame); all 5 are only called from view_model.rs's build_*_vm
functions (plus tests) — a file-placement mismatch, not a runtime violation.

Relocates all 5 into view_model.rs, next to their only real callers
(diff_title/diff_footer -> build_diff_vm, confirm_prompt/confirm_modal_style
-> build_confirm_vm, gist_row_display -> build_list_vm). Pure move: bodies
and existing tests (tui/tests.rs) are unchanged. diff_footer wasn't named in
the issue's original grep, but has the identical shape (only ever called
from build_diff_vm + tests, calls real AppState methods) — included to
actually reach the issue's stated end state of zero AppState business-method
calls in render.rs, not just move the 4 named functions.

Closes the one genuinely live gap: the real palette-paint function used to
read state.palette().map(|p| p.query.clone()) directly to get the live
command-palette query text. PaletteVm gains a query: TextInput field
(TextInput, not String, so input_line can still draw the cursor),
populated in build_palette_vm; render_palette_vm now reads palette.query.

After this, render.rs calls zero AppState business methods — only field
reads (state.theme, state.status, etc.) remain, which is the paint = theme/
layout seam the architecture doc describes.

Not touched (pre-existing, same-shaped gap, flagged for a follow-up issue
rather than expanding scope here): render.rs still reads
state.local_filter_query / state.filter_query directly for
ListFooterVm::Filtering instead of routing through a VM field.

Closes #289
